backend/api-classify-ticket/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
